+++ src/main/java/org/gatodev/arcadiaclinica/repository/medical_services/IMedicalServiceRepository.java
@@ -1,7 +1,7 @@
 package org.gatodev.arcadiaclinica.repository.medical_services;
 
-import org.gatodev.arcadiaclinica.entity.medical_services.MedicalService;
+import org.gatodev.arcadiaclinica.entity.medical.MedicalService;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IMedicalServiceRepository extends JpaRepository<MedicalService, Integer> {
+public interface IMedicalServiceRepository extends JpaRepository<MedicalService, Long> {
 }
