+++ src/main/java/org/gatodev/arcadiaclinica/repository/medical_services/IMedicalSpecialtyRepository.java
@@ -1,7 +1,7 @@
 package org.gatodev.arcadiaclinica.repository.medical_services;
 
-import org.gatodev.arcadiaclinica.entity.medical_services.MedicalSpecialty;
+import org.gatodev.arcadiaclinica.entity.medical.MedicalSpecialty;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IMedicalSpecialtyRepository extends JpaRepository<MedicalSpecialty, Integer> {
+public interface IMedicalSpecialtyRepository extends JpaRepository<MedicalSpecialty, Long> {
 }
