+++ src/main/java/org/gatodev/arcadiaclinica/repository/medical_services/IMedicalTypeServiceRepository.java
@@ -1,9 +1,9 @@
 package org.gatodev.arcadiaclinica.repository.medical_services;
 
-import org.gatodev.arcadiaclinica.entity.medical_services.MedicalTypeService;
+import org.gatodev.arcadiaclinica.entity.medical.MedicalTypeService;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IMedicalTypeServiceRepository extends JpaRepository<MedicalTypeService, Integer> {
+public interface IMedicalTypeServiceRepository extends JpaRepository<MedicalTypeService, Long> {
 }
